+++ components/atom/popover/src/index.js
@@ -1,15 +1,9 @@
-import {
-  Children,
-  cloneElement,
-  useEffect,
-  useState,
-  useRef,
-  forwardRef
-} from 'react'
+import {useRef, forwardRef} from 'react'
 import PropTypes from 'prop-types'
-import useControlledState from '@s-ui/react-hooks/lib/useControlledState'
 import loadable from '@loadable/component'
+import useControlledState from '@s-ui/react-hooks/lib/useControlledState'
 import {PLACEMENTS, TRIGGERS} from './config'
+import PopoverExtendChildren from './PopoverExtendChildren'
 
 const BASE_CLASS = 'sui-AtomPopover'
 const CLASS_INNER = `${BASE_CLASS}-inner`
@@ -18,7 +12,7 @@ const DEFAULT_OFFSET = 'auto,4px'
 const DEFAULT_TRIGGER = TRIGGERS.LEGACY
 const DEFAULT_DELAY = 0
 
-// https://github.com/reactstrap/reactstrap/blob/7.1.0/src/TooltipPopoverWrapper.js
+// https://github.com/reactstrap/reactstrap/blob/8.9.0/src/Popover.js
 const Popover = loadable(() => import('reactstrap/lib/Popover'), {ssr: true})
 
 const AtomPopover = forwardRef(
@@ -38,76 +32,43 @@ const AtomPopover = forwardRef(
     outRef
   ) => {
     const targetRef = useRef()
-    const [ready, setReady] = useState()
     const [isVisibleState, setIsVisibleState] = useControlledState(
       isVisible,
       defaultIsVisible
     )
-    useEffect(() => {
-      setReady(targetRef.current)
-    }, [targetRef, setReady])
-
-    const extendChildren = () => {
-      const onClick = (onClickHandler = () => null) => ev => {
-        onOpen()
-        setIsVisibleState(true)
-        onClickHandler(ev)
-      }
-      const childrenOnly =
-        typeof children === 'string'
-          ? [<span key={1}>{children}</span>]
-          : Children.only(children)
-      const response = Children.map(childrenOnly, (child, key) => {
-        const attrs = {
-          onClick: onClick(child.props.onClick)
-        }
-        attrs.ref = node => {
-          ;[child.ref, targetRef].forEach(ref => {
-            if (typeof ref === 'function') {
-              ref(node)
-            } else if (ref !== null) {
-              ref.current = node
-            }
-          })
-        }
-        attrs.key = key
-        return cloneElement(child, attrs)
-      })
-      return response
-    }
 
     const handleToggle = ev => {
       setIsVisibleState(!isVisibleState)
       onClose(ev)
     }
-
     return (
       <>
-        {extendChildren()}
-        {isVisibleState && ready && (
-          <Popover
-            className={BASE_CLASS}
-            delay={DEFAULT_DELAY}
-            innerClassName={CLASS_INNER}
-            hideArrow={hideArrow}
-            arrowClassName={`${BASE_CLASS}-arrow`}
-            isOpen={isVisibleState}
-            offset={DEFAULT_OFFSET}
-            placement={placement}
-            placementPrefix={PREFIX_PLACEMENT}
-            target={targetRef.current}
-            toggle={handleToggle}
-            trigger={Array.isArray(trigger) ? trigger.join(' ') : trigger}
-            innerRef={outRef}
-          >
-            {closeIcon && (
-              <div className={`${BASE_CLASS}-closeIcon`} onClick={handleToggle}>
-                {closeIcon}
-              </div>
-            )}
-            {content}
-          </Popover>
-        )}
+        <PopoverExtendChildren ref={targetRef} onToggle={handleToggle}>
+          {children}
+        </PopoverExtendChildren>
+        <Popover
+          trigger={Array.isArray(trigger) ? trigger.join(' ') : trigger}
+          isOpen={isVisibleState}
+          toggle={handleToggle}
+          target={targetRef}
+          className={BASE_CLASS}
+          popperClassName="popperClassName"
+          innerClassName={CLASS_INNER}
+          arrowClassName={`${BASE_CLASS}-arrow`}
+          innerRef={outRef}
+          hideArrow={hideArrow}
+          placementPrefix={PREFIX_PLACEMENT}
+          delay={DEFAULT_DELAY}
+          placement={placement}
+          offset={DEFAULT_OFFSET}
+        >
+          {closeIcon && (
+            <div className={`${BASE_CLASS}-closeIcon`} onClick={handleToggle}>
+              {closeIcon}
+            </div>
+          )}
+          {content}
+        </Popover>
       </>
     )
   }
